Update tutorial.m
As described in your paper (Corcoran et al., 2017), Fw corresponds to the frequency width, a function of the frequency resolution. An Fw value would not imply a frequency resolution of ~0.24 Hz as this is a function of the sampling rate and window. Hopefully my edit here makes things more clear for others. Thank you for the wonderful package!